pyproject.toml
[tool.poetry]
name = "oss_data_quality_frameworks_compare"
version = "0.1.0"
description = "Project with sample code to compare 3 popular DQ pyspark compatible frameworks"
authors = ["Sidharth Bolar <Your Email>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9.6"
pyspark = "^3.0.0"
quinn = "^0.5.0"
sklearn = "^0.0.post5"
scikit-learn = "^1.2.2"
pydeequ = "^1.0.1"
soda-core = "^3.0.41"
soda-core-spark-df = "^3.0.41"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
chispa = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"